efficient.rounding {AlgDesign}	R Documentation

Efficient Rounding

Description

A vector of proportions is efficiently rounded to integers.

Usage

efficient.rounding(proportions,n,random=TRUE)

Arguments

proportions
A vector of proportions.

n
The sum of the resulting integers.

random
If TRUE, ties will be broken at random, otherwise the first of the tied values will be used.

Details

This function implements an efficient rounding procedure to round approximate theory designs into replicated integer approximations.

Value

A vector of replications summing to n.

Author(s)

Bob Wheeler bwheelerg@gmail.com

Please cite this program as follows:

Wheeler, R.E. (2004). efficient.rounding. AlgDesign. The R project for statistical computing https://www.r-project.org/

References

Pulkesheim, F. and Rieder, S. (1992). Efficient rounding of approximate designs. Biometrika. 79-4. 763-770.

[Package AlgDesign version 1.1-7.3.1 Index]