eval.blockdesign {AlgDesign}	R Documentation

Evaluates a blocked design.

Description

A blocked design is evaluated.

Usage

eval.blockdesign(frml,design,blocksizes,rho=1,confounding=FALSE,center=FALSE)

Arguments

frml
The formula used to create the blocked design.

design
The blocked design, which may be the design output by optBlock().

blocksizes
A vector of blocksizes for the design.

rho
A vector, giving the ratios of whole to within variance components.

confounding
If confounding=TRUE, the confounding matrix will be output.

center
If TRUE, numeric variables will be centered before frml is applied.

Value

confounding
A matrix based on the design matrix in which the within block variables have been centered about their block means. The columns of this matrix which give the regression coefficients of each variable regressed on the others. If C is the confounding matrix, then -XC is a matrix of residuals of the variables regressed on the other variables.

determinant.all.terms.within.terms.centered
(det(M)^(1/k), where M=X'X/N and X is the model expanded N x k design matrix in which the within block variables have been centered about the grand mean.

within.block.efficiencies
The determinant criterion blocking efficiencies for the range of rho's input. A high efficiency indicates that there is little intrablock information to be recovered in the analysis.

block.centered.properties
A matrix with four rows. The columns correspond to constant, whole block terms and within block terms:

The degrees of freedom for terms in the expanded model.

The determinant of the block centered within block terms.

The geometric mean of the block centered variances.

The geometric mean of the ratio of centered to block centered variances.

Author(s)

Bob Wheeler bwheelerg@gmail.com

Please cite this program as follows:

Wheeler, R.E. (2004). eval.blockdesign. AlgDesign. The R project for statistical computing https://www.r-project.org/

[Package AlgDesign version 1.1-7.3.1 Index]