eval.design {AlgDesign}	R Documentation

Evaluates a design.

Description

A design is evaluated.

Usage

eval.design(frml,design,confounding=FALSE,variances=TRUE,center=FALSE,X=NULL)

Arguments

frml
The formula used to create the design.

design
The design, which may be the design part of the output of optFederov().

confounding
If confounding=TRUE, the confounding patterns will be shown.

variances
If TRUE, the variances each term will be output.

center
If TRUE, numeric variables will be centered before frml is applied.

X
X is either the matrix describing the prediction space for I or for G, the the candidate set from which the design was chosen. They are often the same.

Value

confounding
A matrix. The columns of which give the regression coefficients of each variable regressed on the others. If C is the confounding matrix, then -ZC is a matrix of residuals of the variables regressed on the other variables.

determinant
(det(M/N)^(1/k), where M=Z'Z/N, and Z is the model expanded N x k design matrix.

A
The average coefficient variance: trace(Mi)/k, where Mi is the inverse of M.

I
The average prediction variance over X, which can be shown to be trace((X'X*Mi)/N.

Ge
The minimax normalized variance over X, expressed as an efficiency with respect to the optimal approximate theory design. It is defined as k/max(d), where max(d) is the maximum normalized variance over X – i.e. the max of x'(Mi)x, over all rows x' of X.

Dea
A lower bound on D efficiency for approximate theory designs. It is equal to exp(1-1/Ge).

diagonality
The diagonality of the design, excluding the constant, if any. Diagonality is defined as (|M1|/prod(diag(M1)))^(1/k), where M1 is M with first column and row deleted when there is a constant.

gmean.variances
The geometric mean of the coefficient variances.

Note

I, Ge and Dea are calculated only when X is input.

Author(s)

Bob Wheeler bwheelerg@gmail.com

Please cite this program as follows:

Wheeler, R.E. (2004). eval.design. AlgDesign. The R project for statistical computing https://www.r-project.org/

[Package AlgDesign version 1.1-7.3.1 Index]