model.matrix.formula {AlgDesign}	R Documentation

Builds a model matrix

Description

Produces a model matrix using expand.formula()

Usage

## S3 method for class 'formula'
model.matrix(frml,data,...)

Arguments

frml
A formula.

data
A data.frame

...
Additional arguments passed to model.matrix.default()

Details

This is a method function signaled by a formula as the first argument. It causes the formula to be translated by expand.formula() before calling model.matrix.default().

Value

A matix.

Author(s)

Bob Wheeler bwheelerg@gmail.com

Please cite this program as follows:

Wheeler, R.E. (2004). model.matrix.formula. AlgDesign. The R project for statistical computing https://www.r-project.org/

[Package AlgDesign version 1.1-7.3.1 Index]