model.matrix.terms {AlgDesign}	R Documentation

Internal model.matrix.terms function

Description

defines a model.matrix function for the terms class

Usage

## S3 method for class 'terms'
model.matrix(...) 

Details

This is not to be called by the user. It fixes a bug whose solution was suggested by Bill Dunlap, TIBCO Software Inc.

[Package AlgDesign version 1.1-7.3.1 Index]