TGTable3 {AlgDesign}	R Documentation

Trinca Gilmour Table 3

Description

A blocked experimental design from Trinca and Gilmour (2001).

A data.frame giving four three level variable in five blocks. There are 45 trials in all.

Usage

	data(TGTable3)

Source

Trinca, L.A. and Gilmour, S.G. (2001). Multistratum response surface designs. Technometrics. 43-1. 25-33.

[Package AlgDesign version 1.1-7.3.1 Index]