TGTable5 {AlgDesign}	R Documentation

Trinca Gilmour Table 5

Description

A blocked experimental design from Trinca and Gilmour (2001).

A matrix giving one whole plot and four within plot variables, in 21 blocks of 2.

Usage

	data(TGTable5)

Source

Trinca, L.A. and Gilmour, S.G. (2001). Multistratum response surface designs. Technometrics. 43-1. 25-33.

[Package AlgDesign version 1.1-7.3.1 Index]