GVTable1 {AlgDesign}	R Documentation

Goos Vandebroek Table 1

Description

A blocked experimental design from Goos and Vandebroek (2003).

A matrix giving one whole plot and four within plot variables, in 21 blocks of 2. The same as TGTable5, but produced by Goos Vandoebroek algorithm.

Usage

	data(GVTable1)

Source

Goos, P. and Vandebroek, M. (2003). D-optimal split-plot designs with given numbers and sizes of whole plots. Technometrics. 45-3. 235-245.

[Package AlgDesign version 1.1-7.3.1 Index]